% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sum_stats_class_and_methods.r
\docType{class}
\name{area.above.threshold-class}
\alias{area.above.threshold-class}
\alias{area.above.threshold}
\title{Area above a threshold}
\description{
Calculates the area of water above a threshold temperature
}
