% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stats_class_and_methods.r
\docType{class}
\name{area.threshold-class}
\alias{area.threshold-class}
\alias{area.threshold}
\title{Threshold area}
\arguments{
\item{threshold}{Critical threshold value - a numeric of length 1}

\item{above}{Logical value - TRUE indicates that we wish to test for values above the threshold. FALSE below.}
}
\value{
\code{threshold.area} returns a tibble  corresponding to the integrated area above (or below) the threshold value.
}
\description{
Returns the area above or below a threshold
}
