% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_class_and_methods.r
\name{bounds.to.indices}
\alias{bounds.to.indices}
\title{Convert z-layer bounds to layer indices}
\usage{
bounds.to.indices(z.range, layer.top, layer.bottom)
}
\arguments{
\item{z.range}{the vertical range}

\item{layer.top}{a vector defining the depths of the tops of the layers}

\item{layer.bottom}{a vector defining the depths of the bottoms of the layers}
}
\value{

}
\description{
Converts a vertical range (two elements) to the indices of the vertical levels defined by their
upper and lower bounds. A layer is included if the bound includes more than 50% of it
}
