% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_class_and_methods.r
\name{chunk.data.source}
\alias{chunk.data.source}
\title{Chunk data source}
\usage{
chunk.data.source(obj, n = 1)
}
\arguments{
\item{obj}{Data source object}

\item{n}{Number of chunks into which to split the object}
}
\value{
A data source object configured for use as chunks
}
\description{
Splits an existing data source object into equally sized chunks
}
