% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stats_class_and_methods.r
\docType{class}
\name{custom.stat}
\alias{custom.stat}
\title{Custom Stat}
\description{
Applies a custom function on a pixel-by-pixel basis, potentially integrating it to a scalar
response variable. The responsibility for the implementation of this model lies with the user - the only
check made here is that the function returns a tibble with columns "resultName", "field" and "value"
}
