% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_class_and_methods.r
\docType{class}
\name{data.source.chunk-class}
\alias{data.source.chunk-class}
\alias{data.source.chunk}
\title{Data source chunk}
\description{
A data chunk based on a subset of a full data.source object
}
\section{Slots}{

\describe{
\item{\code{chunk.id}}{numeric. Unique identifier}
}}

