% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_Common_elements.r
\name{date_to_ym}
\alias{date_to_ym}
\title{Date to year month}
\usage{
date_to_ym(d)
}
\arguments{
\item{d}{Vector of dates to convert, as Date classes}
}
\value{
String of format yyyy-mm
}
\description{
Converts a date to a standardised year-month representation. Handy for standardising time where there is
a risk that the days of the month may be different between data sources.
}
