% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Workload_management.r
\name{get.subdomain.dir}
\alias{get.subdomain.dir}
\title{Get Subdomain directory}
\usage{
get.subdomain.dir(cfg, sp)
}
\arguments{
\item{cfg}{A PredEng config object}

\item{sp}{A Spatial configuration object}
}
\value{
Path to the scratch directory
}
\description{
Gets the scratch directory in which to place any subdomain processing
}
