% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_class_and_methods.r
\name{get.vertical.levels}
\alias{get.vertical.levels}
\title{Get vertical levels}
\usage{
get.vertical.levels(this.cfg, this.datasrc, f)
}
\arguments{
\item{this.cfg}{}

\item{this.datasrc}{}

\item{f}{}
}
\value{

}
\description{
Gets vertical levels from an arbitrary file
}
