% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stats_class_and_methods.r
\docType{class}
\name{habitat}
\alias{habitat}
\title{Habitat model}
\description{
Applies a habitat suitability model to calculate the habitat suitability on a pixel-by-pixel
basis and then integrates it
}
