% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sum_stats_class_and_methods.r
\docType{class}
\name{habitat.suitability-class}
\alias{habitat.suitability-class}
\alias{habitat.suitability}
\title{Habitat suitability model}
\description{
Applies a habitat suitability model
}
