% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sum_stats_class_and_methods.r
\docType{class}
\name{isoline.lat-class}
\alias{isoline.lat-class}
\alias{isoline.lat}
\title{Northward extent of an isolone}
\description{
Calculates the average latitudinal position extent of an isoline
}
