% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Workload_management.r
\name{job_management}
\alias{job_management}
\alias{partition.workload}
\alias{get.this.src}
\alias{get.this.sp}
\alias{get.this.cfgs}
\title{Manage job distribution}
\usage{
partition.workload(obj, src.slot, ensmean = FALSE,
  partition.by.space = !obj@use.global.ROI)

get.this.src(fname, cfg.idx, obj)

get.this.sp(fname, cfg.idx, obj)

get.this.cfgs(fname)
}
\arguments{
\item{obj}{A PredEng.config object}

\item{src.slot}{Vector containg the name of the slot over which to partition 
the workload. In the case where this is set to "SumStats", work is partitioned
over all relevant model data slots}

\item{ensmean}{Treat this as an ensemble mean?}

\item{partition.by.space}{Should the work be split up by spatial subdomain as well as data.src? The default
behaviour is to follow the use.global.ROI slot in obj i.e. when use.global.ROI is TRUE, we don't want to 
partition by space. However, in some cases (e.g. calculation of statistics) it is useful to have this 
behaviour anyway.}

\item{fname}{Filename (incluiding path) to read / write job configuration to/from}

\item{cfg.idx}{Configuration index, indicating which job configration to extract}
}
\description{
Manage job distribution
}
