% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_Common_elements.r
\docType{data}
\name{ll.crs}
\alias{ll.crs}
\title{Predictability Engine Common Elements
==========================================================================}
\format{
An object of class \code{CRS} of length 1.
}
\usage{
ll.crs
}
\description{
by Mark R Payne
DTU-Aqua, Charlottenlund, Denmark
http://www.staff.dtu.dk/mpay
}
\details{
Fri May 20 09:55:40 2016

Defines a set of common-baseline elements for use across all pieces of code
in this codebase
Default CRS
}
\keyword{datasets}
