% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_Common_elements.r
\name{month_fns}
\alias{month_fns}
\alias{date_to_ym}
\alias{month_diff}
\title{Month handling functions month}
\usage{
date_to_ym(d)

month_diff(t1, t2)
}
\arguments{
\item{d}{Vector of dates to convert, as Date classes}

\item{t1}{}

\item{t2}{}
}
\value{
String of format yyyy-mm
}
\description{
Works with dates in terms of their months only, ignoring days. Handy for standardising time where there is
a risk that the days of the month may be different between data sources.
}
