% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stats_class_and_methods.r
\docType{class}
\name{pass.through-class}
\alias{pass.through-class}
\alias{pass.through}
\title{Pass-through statistic}
\description{
Returns the field that was supplied. This can be useful, for example, if we
want to look at anomaly correlation coefficients and just want to pass the 
value straight through into the statistics processing and skill metric system
}
