% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PE.cfg}
\alias{PE.cfg}
\title{Default configuration options for PredEng package}
\format{
A list of configuration variables:
\describe{
  \item{PE.cfg}{Options}
}
}
\source{
Self generated
}
\usage{
PE.cfg
}
\description{
Sets the configuration options for the PredEng package.
}
\keyword{datasets}
