% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.r
\name{PE.current.version}
\alias{PE.current.version}
\title{Get current git version}
\usage{
PE.current.version()
}
\value{
String detailing the current version of the git repository
}
\description{
Get current git version
}
