% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Results_management.R
\name{PE.db.tbl}
\alias{PE.db.tbl}
\title{Parallel-safe tbl constructor}
\usage{
PE.db.tbl(src, from, silent = TRUE)
}
\arguments{
\item{src}{}

\item{from}{}
}
\value{

}
\description{
Creates a tbl object in a way that is concurrency safe
}
