% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_source_class_and_methods.r
\name{PE.get.datasrc}
\alias{PE.get.datasrc}
\title{Extract data.source objecct}
\usage{
PE.get.datasrc(object, this.srcType, this.srcName)
}
\arguments{
\item{object}{PredEng.config object}

\item{this.srcType}{Type of the data.source to match}

\item{this.srcName}{Name of the data.source to match}
}
\value{
data.source object
}
\description{
Extract the correct datasource object from the configuration specifying the
source type and source name.
}
