% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_Common_elements.r
\name{PE.global.sf}
\alias{PE.global.sf}
\title{Get a global ROI as simple features object}
\usage{
PE.global.sf(this.pcfg)
}
\arguments{
\item{this.pcfg}{PCFG, containing a valid global.ROI object}
}
\value{

}
\description{
Get a global ROI as simple features object
}
