% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Config_class_and_methods.r
\name{PE.load.config}
\alias{PE.load.config}
\title{Import PredEng configuration}
\usage{
PE.load.config()
}
\value{
PredEng.config object
}
\description{
Loads the currently configured PredEng.config object. Configuration is defined according to the
configuration file linked into the project directory.
}
