% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Workload_management.r
\name{PE.progress}
\alias{PE.progress}
\title{PredEng Progress bar}
\usage{
PE.progress(n)
}
\arguments{
\item{n}{Number of iterations}
}
\value{
A progress bar object
}
\description{
Creates a standardised progress bar across all of the Predictability Engine
}
