% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common_configuration.r
\name{PE.scratch.path}
\alias{PE.scratch.path}
\title{Get Standard Filenames}
\usage{
PE.scratch.path(pcfg, type)
}
\arguments{
\item{pcfg}{PE.cfg object}

\item{type}{Type of filename  to retrieve}
}
\value{
The path to the particular file
}
\description{
Get Standard Filenames
}
