% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PElst.r
\docType{class}
\name{PElst-class}
\alias{PElst-class}
\title{PElst class}
\description{
Provides a list-like class that can be used to control the contents more strictly. Gratutiously stolen
from FLR's FLlst concept.
}
