% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Config_class_and_methods.r
\name{plot,PredEng.config,missing-method}
\alias{plot,PredEng.config,missing-method}
\title{Visualise project region of interest}
\usage{
\S4method{plot}{PredEng.config,missing}(x, y, ...)
}
\description{
Visualise project region of interest
}
