% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Config_class_and_methods.r
\docType{class}
\name{PredEng.config-class}
\alias{PredEng.config-class}
\alias{PredEng.config}
\title{PredEng Project configuration class}
\description{
PredEng Project configuration class
}
\details{
Extraction from the data sets can be performed either at specific geographical locations
 (for all  time points), at specific times (for all locations), or at specific points in space-time. 
 This extraction process is controlled by the extraction slot, which consists of a list of vectors, 
 as follows
\itemize{
  \item  A vector of dates on which to extract rasters e.g. to make spatial forecast maps
  \item An SpatialPointsDataFrame object from the sp package, indicating the points in
space and time that should be extracted
} 
Different extraction modes can be combined in the same list.
}
\section{Slots}{

\describe{
\item{\code{project.name}}{Name of the configuration}

\item{\code{Observations}}{A data.source object defining the observational dataset to include.}

\item{\code{Decadal}}{A list of GCM objects defining the decadal forecast systems to be analysed}

\item{\code{NMME}}{A list of GCM objects defining the NMME models to be analysed}

\item{\code{CMIP5}}{A list of data source objects to configure extraction from the CMIP5 ensemble.}

\item{\code{obs.only}}{Only process observations. Useful for cases where we don't want to process everything.}

\item{\code{persistence.leads}}{A vector (in months) of lead times at which to generate persistence forcasts}

\item{\code{spatial.polygons}}{A sf data.frame defining the spatial domains over which  to operate}

\item{\code{statistics}}{PredEng.list of statistics to apply over each spatial area}

\item{\code{pt.extraction}}{A tibble defining the points to extract, with three columns 1) table - the database table to do 
the extraction from 2) filter - a character string defining a dplyr::filter command to subset the set of fields
in the table - extraction only tables place from these 3) points - a st_sf data frame defining the points at which
to extract and the date time ("date" column). Other metadata is carried forward into the output database}

\item{\code{calibrationMethods}}{A vector listing the calibration methods to use. The full list of supported values is 
stored in PE.cfg$validity$calibrationMethods}

\item{\code{MOI}}{The months of interest (a vector of integers between 1 and 12 inclusive). Currently only support one month
but this can be extened in the future}

\item{\code{average.months}}{Should we average over the months of interest? Currently redundant as we do not support multiple
month extraction, but may come back in the future}

\item{\code{vert.range}}{The vertical range, in m, over which to average. NA indicates no vertical averaging (for 
2D fields) or to use the surface layer (for 3D fields).}

\item{\code{clim.years}}{The years to include in the climatology and analysis of hindcast skill (vector of 
integers)}

\item{\code{comp.years}}{The years over which to make comparisons between observations and models}

\item{\code{landmask}}{The file to use as the basis for masking out land [optional]. If null, all pixels are used.}

\item{\code{scratch.dir}}{Directory in which processed files are to be stored}

\item{\code{global.ROI}}{Extent object defining the global ROI to work with.}

\item{\code{global.res}}{Resolution of the analysis to be applied globally}

\item{\code{retain.realizations}}{Is there interest in retaining the individual realisations from each model, or
should we go straight to the realisation means (as a way of saving disk space and simiplifying the
processing)?}

\item{\code{package.version}}{Details of the package version that this object was built with. Written automatically by the set.configuration() function}
}}

