% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_class_and_methods.r
\docType{class}
\name{PredEng.source-class}
\alias{PredEng.source-class}
\alias{PredEng.source}
\title{Metadata class for describing data sources compatabile with PredEng}
\description{
Metadata class for describing data sources compatabile with PredEng
}
\section{Slots}{

\describe{
\item{\code{name}}{Identifier of data source}

\item{\code{type}}{Type of data source that this object corresponds to - valid 
options are "Decadal", "NMME", "CMIP","Obs", and eventually "C3S"}
}}

