% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Results_management.R
\name{raster.list.mean}
\alias{raster.list.mean}
\title{Raster List functions}
\usage{
raster.list.mean(l)
}
\arguments{
\item{l}{List of raster layers}
}
\value{

}
\description{
Helper functions to work with lists of rasters
}
