% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Field_functions.r
\name{rasterList}
\alias{rasterList}
\alias{raster.list.mean}
\alias{raster.list.sd}
\title{Raster List functions}
\usage{
raster.list.mean(l)

raster.list.sd(l)
}
\arguments{
\item{l}{List of raster layers}
}
\value{

}
\description{
Helper functions to work with lists of rasters
}
