% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.r
\name{sfpolygon.from.extent}
\alias{sfpolygon.from.extent}
\title{Create sf polygon from raster extent object}
\usage{
sfpolygon.from.extent(ext)
}
\arguments{
\item{ext}{raster extent object}
}
\value{
sf POLYGON object
}
\description{
Create sf polygon from raster extent object
}
