% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stats_class_and_methods.r
\docType{class}
\name{spatial.mean-class}
\alias{spatial.mean-class}
\alias{spatial.mean}
\title{Average temperature within an ROI}
\description{
Calculates the average temperature within a region of interest using
area-weighting
}
