% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stats_class_and_methods.r
\docType{class}
\name{stat-class}
\alias{stat-class}
\alias{stat}
\title{Statistics class}
\arguments{
\item{name}{Name of the statistics class}

\item{use.realmeans}{Indicates whether to use the mean of the individ realisations or
the realisation values themselves. This slot essentially acts as a flag telling
the script whether it wants 2D (lat-lon) or 3D (lat-lon-realization) data.}

\item{use.anomalies}{Should the statistic be calculated on the basis of
anomalies only or should the full field be used?}
}
\description{
Statistics class
}
