% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stats_class_and_methods.r
\docType{class}
\name{stat}
\alias{stat}
\title{Statistics class}
\arguments{
\item{name}{Name of the statistics class. Cannot contain spaces, underscores or dots. Must be less than 20 char}

\item{desc}{Description of the statistics class.}

\item{apply.to.realisations}{Should we apply the statistic to all realisations? Realmeans, Ensmean and Observations (including calibrated observations) are always calculated.}

\item{use.globalROI}{Indicates whether the stat should also be calculated on
global basis, in addition to those specified by this.stat@spatial.polygons slot.}

\item{spatial.polygons}{The names of spatial polygons to run this metric with. Setting this to 
NULL (the default) will use all available polygons. NA will use none.}

\item{retain.field}{Should fields generated by the metric be retained in the output? 
This can be expensive in terms of storage, so it can be turned off here.}

\item{do.field.metrics}{Should field metrics be calculated for this particular stat. Defaults to FALSE. These
can be computationally expensive, so need to be opted into.}
}
\description{
Statistics class
}
