% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sum_stats_class_and_methods.r
\docType{class}
\name{sum.stat-class}
\alias{sum.stat-class}
\title{Summary statistics class}
\description{
Summary statistics class
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of the Summary statistics class}

\item{\code{data.type}}{Indicates whether to use the individ. Valid values are 
"realizations" and "means". This slot essentially acts as a flag telling
the script whether it wants 2D (lat-lon) or 3D (lat-lon-realization) data.}

\item{\code{use.anomalies}}{Should the summary statistic be calculated on the basis of
anomalies only or should the full field be used?}
}}

