% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stats_class_and_methods.r
\docType{class}
\name{threshold.area-class}
\alias{threshold.area-class}
\alias{threshold.area}
\title{Threshold Area}
\value{
Tibble
}
\description{
Calculates the area of water above (or below) a threshold value
}
