% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_source_class_and_methods.r
\name{timebounds.to.time}
\alias{timebounds.to.time}
\title{Timebounds to Time var}
\usage{
timebounds.to.time(this.obj, f)
}
\arguments{
\item{this.obj}{A data source object}

\item{f}{The path of the variable to be modified}
}
\description{
In cases where a netcdf file has time boundaries associated with it, it can
be useful to copy one of these values into the time variable, to facilitate
easier processing.
}
