% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_source_class_and_methods.r
\name{verticalLayers}
\alias{verticalLayers}
\alias{bounds.to.indices}
\alias{midpoints.to.indices}
\title{Convert vertical coordinates to layer indices}
\usage{
bounds.to.indices(z.range, layer.top, layer.bottom)

midpoints.to.indices(z.range, midpoints)
}
\arguments{
\item{z.range}{the vertical range}

\item{layer.top}{a vector defining the depths of the tops of the layers}

\item{layer.bottom}{a vector defining the depths of the bottoms of the layers}

\item{midpoints}{A vector of the midpoints}
}
\value{



}
\description{
bounds.to.indices converts a vertical range (two elements) to the indices of the vertical levels defined by their
upper and lower bounds. A layer is included if the bound includes more than 50% of it
}
