% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stats_class_and_methods.r
\docType{class}
\name{area.above.threshold-class}
\alias{area.above.threshold-class}
\alias{area.above.threshold}
\title{Area above a threshold}
\arguments{
\item{threshold}{Critical threshold value}
}
\description{
Calculates the area of water above a threshold temperature
}
