% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_class_and_methods.r
\docType{class}
\name{data.source-class}
\alias{data.source-class}
\alias{data.source}
\title{Metadata class for describing data from a single model}
\description{
Metadata class for describing data from a single model
}
\section{Slots}{

\describe{
\item{\code{name}}{Identifier of data source}

\item{\code{type}}{Type of data source that this object corresponds to - valid 
options are "Decadal", "NMME", "CMIP","Obs", and eventually "C3S"}

\item{\code{source}}{Directory or URL containing the raw information}

\item{\code{n.chunks}}{Number of chunks into which to split the processing}

\item{\code{var}}{Variable name from which to extract data}

\item{\code{levels}}{Level(s) in the source data set containing the relevant data. A value of NA means that averaging 
take place over all available layers. Also an epic track by Avicii (RIP).}

\item{\code{realizations}}{Character string, naming the realization(s) to use. NA indicates all.}

\item{\code{time.correction}}{A string suitable for use with the "cdo shifttime,x" command string that can be used to 
correcct for time-bound versus time value problems.}

\item{\code{realization.fn}}{A function to extract the realisation ID}

\item{\code{init.fn}}{Function to extract the initialisation dates}

\item{\code{date.fn}}{A function to extract the time stamps for each time step}
}}

