% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_class_and_methods.r
\docType{class}
\name{data.source-class}
\alias{data.source-class}
\alias{data.source}
\title{Metadata class for describing data from a single model}
\description{
Metadata class for describing data from a single model
}
\section{Slots}{

\describe{
\item{\code{name}}{Identifier of data source}

\item{\code{type}}{Type of data source that this object corresponds to - valid 
options are "Decadal", "NMME", "CMIP","Obs", and eventually "C3S"}

\item{\code{n.chunks}}{Numeric. (Max) number of chunks allocated to processing this data source.}

\item{\code{chunk.id}}{Chunk identifier (optional)}

\item{\code{sources}}{A list of directories or URLs containing the raw information. Each item in the list
corresponds to a processing chunk. The number of chunks here might be less than n.chunks, but
should not exceed it}

\item{\code{var}}{Variable name from which to extract data}

\item{\code{realizations}}{Character string, naming the realization(s) to use. NA indicates all.}

\item{\code{use.timebounds}}{Indicates whether to use timebounds instead of "time" variable, when selecting
by time. Valid values are NA (don't use timebounds, use the time variable), 1 (use the lower bound),
2 (use the upper bound) or 3 (use the average value)}

\item{\code{realization.fn}}{A function to extract the realisation ID}

\item{\code{layermids.fn}}{Returns the midpoints of the vertical layers for this data set.}

\item{\code{start.date}}{Function to extract the dates for the start of the forecasts. Note that we define this to
be different to the initialisation dates (when the model is actually initialised). A decadal model 
might be initialised on 1 Nov but we are primarily interested in it's post-January output, and so call 
this an effective start date of 1 Jan. Things are much clearer for seasonal forecast systems, that 
are initialised monthly.}

\item{\code{date.fn}}{A function to extract the time stamps for each time step}
}}

