% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Workload_management.r
\name{job_management}
\alias{job_management}
\alias{partition.workload}
\alias{get.this.src}
\alias{get.this.sp}
\alias{get.this.cfgs}
\title{Manage job distribution}
\usage{
partition.workload(obj, src.slot, data.partition.type = "",
  space.partition = !obj@use.global.ROI)

get.this.src(fname, cfg.idx, obj)

get.this.sp(fname, cfg.idx, obj)

get.this.cfgs(fname)
}
\arguments{
\item{obj}{A PredEng.config object}

\item{src.slot}{Name of the slot over which to partition the workload. In the case where
this is set to "Stats", work is partitioned over all relevant model data slots}

\item{data.partition.type}{"ensmean", "source", or "chunk"}

\item{space.partition}{Should the work be split up by spatial subdomain? The default
behaviour is to follow the use.global.ROI slot in obj i.e. when use.global.ROI is TRUE, we don't want to 
partition by space. However, in some cases (e.g. calculation of statistics) it is useful to have this 
behaviour anyway.}

\item{fname}{Filename (incluiding path) to read / write job configuration to/from}

\item{cfg.idx}{Configuration index, indicating which job configration to extract}
}
\description{
Manage job distribution
}
