% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Workload_management.r
\name{job_management}
\alias{job_management}
\alias{partition.workload}
\alias{configure.src}
\alias{configure.chunk}
\alias{get.cfgs}
\title{Manage job distribution}
\usage{
partition.workload(obj, src.slot = "missing", data.partition.type = "")

configure.src(fname, cfg.idx, obj)

configure.chunk(fname, cfg.idx, obj)

get.cfgs(fname)
}
\arguments{
\item{obj}{A PredEng.config object}

\item{src.slot}{Name of the slot over which to partition the workload. In the case where
this is set to "Stats", work is partitioned over all relevant model data slots}

\item{data.partition.type}{"ensmean", "source", or "chunk"}

\item{fname}{Filename (incluiding path) to read / write job configuration to/from}

\item{cfg.idx}{Configuration index, indicating which job configration to extract}
}
\description{
Manage job distribution
}
