% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PE.cfg}
\alias{PE.cfg}
\title{Default configuration options for PredEng package}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
  \item{PE.cfg}{Options}
}
}
\usage{
PE.cfg
}
\description{
Sets the configuration options for the PredEng package. The variables are as follows:
}
\keyword{datasets}
