% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_Common_elements.r
\name{PE.config.summary}
\alias{PE.config.summary}
\title{Display configuration status}
\usage{
PE.config.summary(...)
}
\arguments{
\item{...}{Relevant paramters to be displayed}
}
\description{
Display configuration status
}
