% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Results_management.R
\name{PE.db}
\alias{PE.db}
\alias{PE.db.connection}
\alias{PE.db.path}
\alias{PE.db.setup}
\alias{PE.db.delete.by.pKey}
\alias{PE.db.delete.by.datasource}
\alias{PE.db.appendTable}
\alias{PE.db.unserialize}
\alias{PE.db.getQuery}
\title{Setup results databases}
\usage{
PE.db.connection(object, table)

PE.db.path(object, table)

PE.db.setup(object)

PE.db.delete.by.pKey(object, table, pKeys, silent = TRUE)

PE.db.delete.by.datasource(
  object,
  table = PE.cfg$db$extract,
  datasrc,
  silent = TRUE
)

PE.db.appendTable(object, table, dat, silent = TRUE, serialize.first = TRUE)

PE.db.unserialize(dat)

PE.db.getQuery(object, table, this.sql, silent = TRUE)
}
\arguments{
\item{object}{PredEng configuration object}

\item{table}{Name of the table to connect to. Each table is stored in a separate file}
}
\value{

}
\description{
Setup results databases
}
\details{
PE.db.appendTable serialises the data column and writes the data to the specified table

PE.db.unserialize  unserialises the data column

PE.db.getQuery performs a concurrency-safe query
}
