% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Results_management.R
\name{PE.db}
\alias{PE.db}
\alias{PE.db.connection}
\alias{PE.db.setup}
\alias{PE.db.clear.datasource}
\title{Setup results database}
\usage{
PE.db.connection(pcfg)

PE.db.setup(pcfg)

PE.db.clear.datasource(this.db, this.datasrc)
}
\arguments{
\item{pcfg}{}
}
\value{

}
\description{
Setup results database
}
