% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Results_management.R
\name{PE.db}
\alias{PE.db}
\alias{PE.db.connection}
\alias{PE.db.setup}
\alias{PE.db.delete.by.pKey}
\alias{PE.db.delete.by.datasource}
\alias{PE.db.calc.realMeans}
\alias{PE.db.appendTable}
\alias{PE.db.unserialize}
\title{Setup results database}
\usage{
PE.db.connection(pcfg)

PE.db.setup(pcfg)

PE.db.delete.by.pKey(pcfg, tbl.name, pKeys)

PE.db.delete.by.datasource(pcfg, tbl.name = PE.cfg$db$extract, datasrc)

PE.db.calc.realMeans(pcfg, this.datasrc)

PE.db.appendTable(dat, pcfg, tbl.name)

PE.db.unserialize(this.dat)
}
\arguments{
\item{pcfg}{}
}
\value{

}
\description{
Setup results database
}
\details{
PE.db.appendTable serialises the data column and writes the data to the specified table

PE.db.unserialize  unserialises the data column
}
