% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Results_management.R
\name{PE.db}
\alias{PE.db}
\alias{PE.db.connection}
\alias{PE.db.setup}
\alias{PE.db.delete.by.pKey}
\alias{PE.db.delete.by.datasource}
\alias{PE.db.appendTable}
\alias{PE.db.unserialize}
\alias{PE.db.getQuery}
\title{Setup results database}
\usage{
PE.db.connection(pcfg, results.db = FALSE)

PE.db.setup(pcfg, results.only = FALSE)

PE.db.delete.by.pKey(pcfg, tbl.name, pKeys, silent = TRUE)

PE.db.delete.by.datasource(
  pcfg,
  tbl.name = PE.cfg$db$extract,
  datasrc,
  silent = TRUE
)

PE.db.appendTable(dat, pcfg, tbl.name, silent = TRUE)

PE.db.unserialize(this.dat)

PE.db.getQuery(pcfg, this.sql, silent = TRUE)
}
\arguments{
\item{pcfg}{}
}
\value{

}
\description{
Setup results database
}
\details{
PE.db.appendTable serialises the data column and writes the data to the specified table

PE.db.unserialize  unserialises the data column

PE.db.getQuery performs a concurrency-safe query
}
