% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stats_class_and_methods.r
\docType{class}
\name{stat}
\alias{stat}
\title{Statistics class}
\arguments{
\item{name}{Name of the statistics class. Cannot contain spaces, underscores or dots. Must be less than 20 char}

\item{desc}{Description of the statistics class.}

\item{realizations}{Tells what sort of data to be considered  1 = Observations for Persistence 
2 = Individual realisations  3 = Realization means 4 = Ensemble means. Observations are always calculated.}

\item{calibration}{Choose the calibration method to base the statistic on. Defaults to NULL, indicating
to use all available calibrations}

\item{use.globalROI}{Indicates whether the stat should also be calculated on global basis, in addition
to those specified by this.stat@spatial.polygons slot.}

\item{spatial.polygons}{The names of spatial polygons to run this metric with. Setting this to 
NULL (the default) will use all available polygons. NA will use none.}

\item{retain.field}{Should fields generated by the metric be retained in the output? 
This can be expensive in terms of storage, so it can be turned off here.}
}
\description{
Statistics class
}
