% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stats_class_and_methods.r
\docType{class}
\name{stat}
\alias{stat}
\title{Statistics class}
\arguments{
\item{name}{Name of the statistics class. Cannot contain spaces, underscores or dots. Must be less than 20 char}

\item{desc}{Description of the statistics class.}

\item{realizations}{Tells what sort of realisations to be considered 1 = Individual realisations 
2 = Realization means 3 = Ensemble means}

\item{calibration}{Choose the calibration method to base the statistic on. Defaults to 
"Mean adjusted".}

\item{use.globalROI}{Indicates whether the stat should be calculated on a local or 
global basis  ie. for each individual spatial polygon (FALSE) or once for the entire global 
ROI (TRUE)}

\item{retain.field}{Should fields generated by the metric be retained in the output? 
This can be expensive in terms of storage, so it can be turned off here.}

\item{skill.metrics}{A character vector listing the skill metrics to be applied for this statistic.}
}
\description{
Statistics class
}
