% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stats_class_and_methods.r
\docType{class}
\name{stat}
\alias{stat}
\title{Statistics class}
\arguments{
\item{name}{Name of the statistics class}

\item{use.realmeans}{Indicates whether to use the mean of the individ realisations or
the realisation values themselves. This slot essentially acts as a flag telling
the script whether it wants 2D (lat-lon) or 3D (lat-lon-realization) data.}

\item{use.full.field}{Should the statistic be calculated on the basis of
anomalies only or should the full field be used?}

\item{use.globally}{Indicates whether the stat should be calculated on a global or local basis 
ie.  once for the entire global ROI (TRUE) or for each individual spatial domain (FALSE)}

\item{retain.field}{Should fields generated by the metric be retained in the output? 
This can be expensive in terms of storage, so it can be turned off here.}

\item{skill.metrics}{A character vector listing the skill metrics to be applied for this statistic.}
}
\description{
Statistics class
}
