% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_class_and_methods.r
\name{test.data.source}
\alias{test.data.source}
\title{Test Data Source}
\usage{
test.data.source(obj, f)
}
\arguments{
\item{obj}{Data.source object}

\item{f}{Filename on which to test the functions}
}
\value{
Displays the outputs generated by the functions
}
\description{
Tests the metadata extraction functions in an object. Useful for testing and developing code.
}
