% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_source_class_and_methods.r
\name{test.data.source}
\alias{test.data.source}
\title{Test Data Source}
\usage{
test.data.source(obj, z.range = c(250, 600), f = "missing")
}
\arguments{
\item{obj}{Data.source object}

\item{z.range}{Range of vertical coordinates over which to test the z2idx function}

\item{f}{Filename on which to test the functions. If missing, defaults to the first file in the data source list}
}
\value{
Displays the outputs generated by the functions
}
\description{
Tests the metadata extraction functions in an object. Useful for testing and developing code.
}
