% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stats_class_and_methods.r
\docType{class}
\name{threshold-class}
\alias{threshold-class}
\alias{threshold}
\title{Threshold}
\arguments{
\item{threshold}{Critical threshold value - a numeric of length 1}

\item{above}{Logical value - TRUE indicates that we wish to test for values above the threshold. FALSE below.}
}
\value{
Raster* object, matching the raster object supplied as an argument
}
\description{
Determines where each pixel sits in relation to a threshold value
}
