% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stats_class_and_methods.r
\docType{class}
\name{threshold-class}
\alias{threshold-class}
\alias{threshold}
\title{Thresholds}
\arguments{
\item{thre}{shold Critical threshold value - a numeric of length 1 or 2}

\item{above}{Logical value - TRUE indicates that we wish to test for values above the threshold. FALSE below. 
Ignored if two thresholds are supplied.}
}
\value{
\code{threshold} returns a tibble containing the Raster* object matching the raster object supplied as an
argument. If two thresholds are supplied, the raster corresponds to that between the two thresholds
}
\description{
Determines where each pixel sits in relation to a threshold value and integrates  over the
area of interest to get the area satisfying the threshold
}
