% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Config_class_and_methods.r
\name{verticalLayers}
\alias{verticalLayers}
\title{Get vertical layers}
\usage{
verticalLayers(cfg, src, ...)
}
\arguments{
\item{cfg}{A PredEng configuration object, with the vert.range slot populated}

\item{src}{A data source object, containing a valid layermids.fn()}

\item{f}{The name of the file from which to extract the layers}
}
\value{
A vector of layer integers corresponding to the vertical range. In cases where the vertical
range limits fall within a layer, the layer is included if it covers more than 50%.
}
\description{
Extracts a list of vertical layers
}
